Move relative filters to another file
