+++ frontend/src/metabase/common/components/FilterPicker/DateFilterPicker/DateFilterPicker.tsx
@@ -1,9 +1,6 @@
 import { useMemo } from "react";
-import {
-  DatePicker,
-  SimpleDatePicker,
-} from "metabase/common/components/DatePicker";
 import type { DatePickerValue } from "metabase/common/components/DatePicker";
+import { DatePicker } from "metabase/common/components/DatePicker";
 import * as Lib from "metabase-lib";
 import { BackButton } from "../BackButton";
 import type { FilterPickerWidgetProps } from "../types";
@@ -56,48 +53,6 @@ export function DateFilterPicker({
         }
         canUseRelativeOffsets
         isNew={isNew}
-        onChange={handleChange}
-      />
-    </div>
-  );
-}
-
-interface SimpleDateFilterPickerProps {
-  query: Lib.Query;
-  stageIndex: number;
-  column: Lib.ColumnMetadata;
-  filter?: Lib.FilterClause;
-  onChange: (filter: Lib.ExpressionClause | undefined) => void;
-}
-
-export function SimpleDateFilterPicker({
-  query,
-  stageIndex,
-  column,
-  filter,
-  onChange,
-}: SimpleDateFilterPickerProps) {
-  const value = useMemo(() => {
-    return filter && getPickerValue(query, stageIndex, filter);
-  }, [query, stageIndex, filter]);
-
-  const availableOperators = useMemo(() => {
-    return getPickerOperators(query, stageIndex, column);
-  }, [query, stageIndex, column]);
-
-  const handleChange = (value: DatePickerValue | undefined) => {
-    if (value) {
-      onChange(getFilterClause(query, stageIndex, column, value));
-    } else {
-      onChange(undefined);
-    }
-  };
-
-  return (
-    <div data-testid="datetime-filter-picker">
-      <SimpleDatePicker
-        value={value}
-        availableOperators={availableOperators}
         onChange={handleChange}
       />
     </div>
