+++ frontend/src/metabase/common/components/FilterPicker/DateFilterPicker/SimpleDateFilterPicker/index.ts
@@ -1,2 +1,1 @@
-export * from "./DateFilterPicker";
 export * from "./SimpleDateFilterPicker";
